mon partie (rue,secteur,cartier) plus

+++ Documents/NetBeansProjects/e-Taxe-Tnb-master/src/main/java/com/fstg/eTaxe/Tnb/service/Impl/SecteurServiceImpl.java
@@ -10,9 +10,11 @@ import com.fstg.eTaxe.Tnb.bean.Secteur;
 import com.fstg.eTaxe.Tnb.dao.SecteurDao;
 import com.fstg.eTaxe.Tnb.service.QuartierService;
 import com.fstg.eTaxe.Tnb.service.SecteurService;
+import java.util.Iterator;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 /**
  *
@@ -25,6 +27,8 @@ public class SecteurServiceImpl implements SecteurService {
     public SecteurDao secteurDao;
     @Autowired
     private QuartierService quartierService;
+//    @Autowired
+//    private SecteurService secteurService;
 
     @Override
     public Secteur findByLibelle(String libelle) {
@@ -37,20 +41,77 @@ public class SecteurServiceImpl implements SecteurService {
     }
     
     @Override
-    public void save(Secteur secteur) {
-        
-        secteurDao.save(secteur);
+    public int saveAndTestExisting(Secteur secteur) {
+        int rep;
+        rep=existByLibelle(secteur.getLibelle());
+        if(rep==1){
+            secteurDao.save(secteur);
+            return 1;
+        }
+     
+            return -1;
     }
-    
+
+   
     @Override
-    public void deleteSecteur(long id) {
-        secteurDao.deleteById(id);
-    }
+    public int deleteSecteur(long id) {
+            Secteur secteur=findByid(id);
+            for(Quartier quartier :secteur.getQuartiers()){
+                if (quartier==null){
+                    secteurDao.deleteById(id);
+                   return 1;
+                }
+            } return -1;
+            
+        }
+       
+       
+    
 
     @Override
     public Secteur WhereDistricExiste(long idQuartier) {
         Quartier quartier =quartierService.findByid(idQuartier);
         return quartier.getSecteur();
     }
+
+    @Override
+    public int exitstsByid(long id) {
+        //throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    if (secteurDao.existsById(id)) {
+            return 1;
+        } else {
+            return 0;
+        }
+    }
+    //already tested
+    @Override
+    public int existByLibelle(String libelle) {
+        //throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+        List<Secteur> secteurs =findAll();
+        for(Secteur secteur:secteurs){
+            if(secteur.getLibelle().equalsIgnoreCase(libelle)){
+                return -1;
+            }
+        }
+                return 1;
+            }
+
+    @Override
+    public Secteur findByid(long id) {
+       // throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.4return 
+       return secteurDao.findByid(id);
+    }
+    @Transactional
+    @Override
+    public void deleteByid(long id) {
+        //throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+        secteurDao.deleteById(id);
+    }
+    }
+
+   
     
-}
+
+  
+    
+
